% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aging.R
\name{aging}
\alias{aging}
\title{Title}
\usage{
aging(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
