% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_ai_prop.R
\name{assign_ai_prop}
\alias{assign_ai_prop}
\title{Title}
\usage{
assign_ai_prop(dat, ai.vec)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
