% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd4_update.R
\name{cd4_update}
\alias{cd4_update}
\title{Title}
\usage{
cd4_update(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
