% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coital_acts.R
\name{coital_acts}
\alias{coital_acts}
\title{Organize coital acts data for calculation of transmission probabilities}
\usage{
coital_acts(dat, at)
}
\arguments{
\item{dat}{master data object}

\item{at}{timestep}
}
\value{
'dat' object with updated coital acts tables for transmission module.
}
\description{
Organize coital acts data for calculation of transmission probabilities
}
\details{
Wrapper function that calls: social_discord_edgelist_df, social_coital_acts, social_role_msm, social_act_type_het, social_condom_use. These functions identify discordonant relationships, organize relevant behavioral attributes, and calcuate number of sex acts per partnership per timestep.
}
\examples{
dat <- social_coital_acts_module(dat,at)
}
