% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_infection_history_coital_acts_table.R
\name{coital_acts_table_fxn}
\alias{coital_acts_table_fxn}
\title{Title}
\usage{
coital_acts_table_fxn(
  model,
  dframe,
  pop,
  ind_inf_time,
  ind_inf_donor,
  ind,
  partners
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
