% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths_tx_time_adjusted.R
\name{deaths_tx_time_adjusted}
\alias{deaths_tx_time_adjusted}
\title{Title}
\usage{
deaths_tx_time_adjusted(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here

}
