% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births.R
\name{evo_arrivals}
\alias{evo_arrivals}
\title{Title}
\usage{
evo_arrivals(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
