% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths.R
\name{evo_departures}
\alias{evo_departures}
\title{Title}
\usage{
evo_departures(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
