% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evonet-package.R
\docType{package}
\name{EvoNet-package}
\alias{EvoNet-package}
\alias{EvoNet}
\title{Integrated Bio-Social Models for HIV Epidemiology}
\description{
\tabular{ll}{
   Package: \tab EvoNet\cr
   Type: \tab Package\cr
   Version: \tab 1.0.0\cr
   Date: \tab TBD\cr
   License: \tab TBD\cr
   LazyLoad: \tab yes\cr
}
}
\details{
EvoNet provides an integrated methodological framework in which to study 
the multi-level biological and behavioral foundations of the population 
dynamics of HIV.
}
\keyword{HIV}
\keyword{evolution}
\keyword{package}
