% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evonet_run.R
\name{evonet_run}
\alias{evonet_run}
\title{Title}
\usage{
evonet_run(
  user_base_params = NULL,
  user_var_param = NULL,
  model_names = "evomodel",
  outpath = getwd(),
  module_list = list(initialize.FUN = initialize_module, plot_nw.FUN =
    plot_network_fxn, aging.FUN = vital_aging_module, testing.FUN =
    social_testing_diagnosis_module, treatment.FUN = social_treatment_module,
    update_vl.FUN = viral_update_gamma, update_cd4.FUN = viral_update_cd4_daily,
    coital_acts.FUN = social_coital_acts_module, trans.FUN = transmission_main_module,
    trans_book.FUN = transmission_bookkeeping_module, trans_cd4.FUN =
    transmission_cd4_module, deaths.FUN = vital_deaths_module, births.FUN =
    vital_births_module,      social_trans.FUN = social_attribute_transition_module,
    summary.FUN = summary_module, resim_nets.FUN = EpiModel::resim_nets, verbose.FUN =
    NULL)
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
