% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoplot.R
\name{evoplot}
\alias{evoplot}
\title{Title}
\usage{
evoplot(
  model,
  outpath = getwd(),
  name = NULL,
  nw_stats = TRUE,
  max_points_rep = NULL,
  popsumm_frequency = NULL,
  variables = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
