% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoplot.R
\name{evoplot_internal}
\alias{evoplot_internal}
\title{Title}
\usage{
evoplot_internal(
  model,
  save = TRUE,
  name = NULL,
  outpath = getwd(),
  variables = NULL,
  main = NULL
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
