% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_incidence_rate.R
\name{incidence_rate_plot}
\alias{incidence_rate_plot}
\title{Title}
\usage{
incidence_rate_plot(model_obj)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
