% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_infection_history_vl_plots.R
\name{infection_history_vl_cd4_plot}
\alias{infection_history_vl_cd4_plot}
\title{Title}
\usage{
infection_history_vl_cd4_plot(ind, vl_df)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
