% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_infection_summary_table.R
\name{infection_summary_table_fxn}
\alias{infection_summary_table_fxn}
\title{Title}
\usage{
infection_summary_table_fxn(pop, partners, dframe, ind)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
