% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_agents.R
\name{initialize_agents}
\alias{initialize_agents}
\title{Initialize agent attributes}
\usage{
initialize_agents(dat, at)
}
\arguments{
\item{dat}{master data object}

\item{at}{timestep}
}
\value{
'dat' object with 'pop'list of attributes for each agent appended.
}
\description{
Initialize agent attributes
}
\details{
Subfunction in 'initialize_module'. 
Creates pop data object which contains the agent attribute values.  The attribute names for each agent is returned by input_parameters_agent_attributes. Default values are NA, other values are calculated by subfunction new_additions_fxn.
}
\examples{
dat  <- initialize_agents(dat,at=1)
}
