% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_evonet_misc.R
\name{initialize_evonet_misc}
\alias{initialize_evonet_misc}
\title{Initialization of EvoNet data objects}
\usage{
initialize_evonet_misc(dat)
}
\arguments{
\item{dat}{master data object}
}
\value{
Returns "data" object with additional EvoNet data objects appended.
}
\description{
Initialization of EvoNet data objects
}
\details{
Performs miscellaneous initialization steps to setup evonet model run:  i) creates 'coital_acts_list', which saves the object discord_coital df (table of relevant agent attributes in determining transmission probability) created each time step in social_coital_act; list filled in summary_misc; ii) creates object 'InfMat' which for each new infections saves time of infections, id of infectee, and id of infector; list filled in transmission_bookkeeping_module; iii) creates the popsumm data object which stores model summary statistics each timestep iv) creates age_list, which stores age data of population at beginning of model run and at quarterly inverals for histogram plots created in plots_popsumm; v) creates partner_list object which track for each agent the time a new relationship starts.
}
\examples{
dat <- initialize_evonet_misc(dat)
}
