% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_infecteds_cd4.R
\name{initialize_infecteds_cd4}
\alias{initialize_infecteds_cd4}
\title{Initialize CD4 values}
\usage{
initialize_infecteds_cd4(dat, at)
}
\arguments{
\item{dat}{master data object}

\item{at}{timestep}
}
\value{
'dat' object with agent attributes for CD4 values calculated
}
\description{
Initialize CD4 values
}
\details{
Subfunction in 'initialize_module'. Fills in initial CD4 and associated values for initial infecteds; must be run after initial_infecteds_cd4 as SPVL values are required.
}
\examples{
dat <- initialize_infecteds_cd4(dat,at=1)
}
