% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_infecteds_vl.R
\name{initialize_infecteds_vl}
\alias{initialize_infecteds_vl}
\title{Initialize viral load and SPVL values}
\usage{
initialize_infecteds_vl(dat, at)
}
\arguments{
\item{dat}{master data object}

\item{at}{timestep}
}
\value{
'dat' object with agent attributes for viral load and SPVL calculated.
}
\description{
Initialize viral load and SPVL values
}
\details{
Subfunction in 'initialize_module'. Calculates initial viral load and associated values for initial infected agents.
}
\examples{
initialize_infecteds_vl(dat,at=1)
}
