% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_module.R
\name{initialize_module}
\alias{initialize_module}
\title{EvoNet model initialization}
\usage{
initialize_module(x, param, init, control, s)
}
\arguments{
\item{x}{An EpiModel object of class netest}

\item{init}{An EpiModel object of class init.net}

\item{control}{An EpiModel object of class control.net}

\item{s}{Simulation number}

\item{y}{A list of EvoNet parameters}
}
\value{
Returns "dat" list, which is master data structure for simulation
}
\description{
EvoNet model initialization
}
\details{
Sets up internal data structures and initial virological and demographic values for the initial population. with the following steps:  i) setups internal evonet data structures pop and popsumm with initialize_agents(); ii) runs EpiModel function initialize.net(); iii) fills in virological and associated values for initial infected with initialize_infecteds_vl(); iv) updates VL values of initial infected based on assumed time of infection with viral_update_aim2(); v) creates data structure (vl_list) to store daily VL values of infected agents; vi) creates optional data structures based on user input with initialize_evonet_misc; vii) calculates summary statistics for initial population with summary_popsumm().
}
