% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_popsumm_dynamics.R
\name{initialize_popsumm_dynamics}
\alias{initialize_popsumm_dynamics}
\title{Initialize 'popsumm' data object}
\usage{
initialize_popsumm_dynamics(dat, at)
}
\arguments{
\item{dat}{master data object}

\item{at}{timestep}
}
\value{
'dat' object with 'popsumm' data object appended.
}
\description{
Initialize 'popsumm' data object
}
\details{
Subfunction in 'initialize_module'. Sets up popsumm list and fills in default value of NA or 0 for all elements; then fills in initial value; popsumm is list of various statistics calculated each timestep to describe network/population/epidemic.
}
\examples{
dat <- initialize_popsumm_dynamics(dat,at=1)
}
