% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_vl_cd4_list.R
\name{initialize_vl_cd4_list}
\alias{initialize_vl_cd4_list}
\title{Title}
\usage{
initialize_vl_cd4_list(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
