% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parameters_age_distribution.R
\name{input_parameters_age_distribution}
\alias{input_parameters_age_distribution}
\title{Title}
\usage{
input_parameters_age_distribution(
  initial_param,
  mort_per_timestep_female,
  pop_growth_rate_timestep,
  baseline_input_exp_growth,
  data_name = "usa_men_18_to_100",
  min_age,
  max_age
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
