% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parameters_agent_attributes.R
\name{input_parameters_agent_attributes}
\alias{input_parameters_agent_attributes}
\title{Title}
\usage{
input_parameters_agent_attributes()
}
\value{
Vector of agent attributes as character string.
}
\description{
Returns vector of individual agent attributes (e.g., 'age', 'sex', 'SPVL') that populate 'pop' data structure. This vector is input to initialize_agents (called in initialize_module) which creates a list with the number of elements equal to the number of agent attributes and the length of each element equal to the size of the initial population.
}
\details{
No function arguments.
}
\examples{
agent_attributes <- input_parameters_agent_attributes()
}
