% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parameters_asmr.R
\name{input_parameters_asmr}
\alias{input_parameters_asmr}
\title{Title}
\usage{
input_parameters_asmr(data_name = "usa_men_18_to_100", min_age, max_age)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
