% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parameters_process.R
\name{input_parameters_process}
\alias{input_parameters_process}
\title{Process raw input parameter list for model run}
\usage{
input_parameters_process(params)
}
\value{
list of parameter values.
}
\description{
Calculates derived parameters, qa/qc check, and converts input parameter list 
into EpiModel object
}
\details{
Additional details here
}
\examples{
evoparams <- input_parameters()
}
