% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parameters_qa_qc.R
\name{input_parameters_qaqc}
\alias{input_parameters_qaqc}
\title{Check parameter values for plausible values.}
\usage{
input_parameters_qaqc(evonet_params)
}
\arguments{
\item{evonet_params}{A list of EvoNet parameters.}
}
\value{
return value here.
}
\description{
Evaluates parameter values for plausible values, returns warning/error when warranted (e.g., initial population size is non-positive). Mostly placeholder currently.
}
\details{
Additional details here
}
\examples{
primary_parameters  <- input_parameters_primary()
cd4_data            <- input_parameters_cd4_data()
evoparams <- c(primary_parameters, cd4_data)
input_parameters_qaqc(evoparams)

}
