% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsim_restart.R
\name{netsim_restart}
\alias{netsim_restart}
\title{Title}
\usage{
netsim_restart(x, param, init, control, type = "initial", dat_input = NULL)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
