% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_output.R
\name{output_plots_fxn}
\alias{output_plots_fxn}
\title{Title}
\usage{
output_plots_fxn(evomodel, plot_index = "all", nsim = 1)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
