% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_hist.R
\name{plot_age_hist}
\alias{plot_age_hist}
\title{Title}
\usage{
plot_age_hist(model)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
