% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aim3_drug_muts.R
\name{plot_aim3_drug_muts}
\alias{plot_aim3_drug_muts}
\title{Title}
\usage{
plot_aim3_drug_muts(model)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
