% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_disc_coital_acts_history.R
\name{plot_disc_coital_acts_history}
\alias{plot_disc_coital_acts_history}
\title{Title}
\usage{
plot_disc_coital_acts_history(model, outpath, sim = 1, name = "evomodel")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
