% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_donors_cd4.R
\name{plot_donors_cd4}
\alias{plot_donors_cd4}
\title{Title}
\usage{
plot_donors_cd4(model)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
