% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_end_epidemic_values.R
\name{plot_end_epidemic_values}
\alias{plot_end_epidemic_values}
\title{Title}
\usage{
plot_end_epidemic_values(model_names, var, nyears)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
