% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heritability_quarters.R
\name{plot_heritability_quarters}
\alias{plot_heritability_quarters}
\title{Title}
\usage{
plot_heritability_quarters(model)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
