% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationship_history.R
\name{plot_relationship_history}
\alias{plot_relationship_history}
\title{Title}
\usage{
plot_relationship_history(model, sim, outpath, name = "evomodel")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
