% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_summary_timeseries.R
\name{plot_summary_timeseries}
\alias{plot_summary_timeseries}
\title{Title}
\usage{
plot_summary_timeseries(model, outpath, name = NULL, nw_stats = TRUE)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
