% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table_parameters.R
\name{plot_table_parameters}
\alias{plot_table_parameters}
\title{Title}
\usage{
plot_table_parameters(params, outpath, open = FALSE)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
