% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_to_removal.R
\name{plot_time_to_removal}
\alias{plot_time_to_removal}
\title{Title}
\usage{
plot_time_to_removal(model, type = "aids")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
