% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vl_trajectories.R
\name{plot_vl_trajectories}
\alias{plot_vl_trajectories}
\title{Title}
\usage{
plot_vl_trajectories(
  model,
  sim = 1,
  Therapy_Type = 1,
  outpath = NULL,
  name = "evomodel"
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
