% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vl_trajectories_aim2.R
\name{plot_vl_trajectories_aim2}
\alias{plot_vl_trajectories_aim2}
\title{Title}
\usage{
plot_vl_trajectories_aim2(model, sim, outpath, name)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
