% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_popsumm_multi_models.R
\name{plots_popsumm_multi_models}
\alias{plots_popsumm_multi_models}
\title{Title}
\usage{
plots_popsumm_multi_models(
  model_paths,
  model_names,
  outpath,
  out_names,
  nw_stats = TRUE,
  max_points_rep = 50
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
