% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_risk_compensation.R
\name{prep_risk_compensation}
\alias{prep_risk_compensation}
\title{PrEP risk compensation module}
\usage{
prep_risk_compensation(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
PrEP risk compensation module parameterized with MSM values
}
\details{
Additional details here
}
\examples{
example function call here
}
