% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_risk_factors.R
\name{prep_risk_factors}
\alias{prep_risk_factors}
\title{PrEP risk factor module}
\usage{
prep_risk_factors(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
module to determine if agent has risk factors for PrEP
}
\details{
Additional details here
}
\examples{
example function call here
}
