% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restart_module.R
\name{restart_module}
\alias{restart_module}
\title{Title}
\usage{
restart_module(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here

}
