% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_group_changes.R
\name{risk_group_changes}
\alias{risk_group_changes}
\title{Title}
\usage{
risk_group_changes(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
