% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_epimodel_control_object.R
\name{setup_epimodel_control_object}
\alias{setup_epimodel_control_object}
\title{Run EpiModel's control.net() with EvoNet parameterization.}
\usage{
setup_epimodel_control_object(evonet_params, module_list)
}
\arguments{
\item{evonet_params}{List of EvoNet parameters.}

\item{module_list}{List of EvoNet functions/modules to simulate epidemic.}
}
\value{
EpiModel 'control' object.
}
\description{
Helper function to fill in and run EpiModel control.net(). An important component is the save.other vector created within the function. This vector of object names specifies the objects that will be returned at the end of the model run (i.e., after netsim called) beyond the default EpiModel objects. Thus, any new output objects created must be specified with the save.other vector.
}
\details{

}
\examples{
evocontrol <- setup_epimodel_control_object(evonet_params = evoparams,
     module_list   = evo_module_list)
}
