% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_2nd_line_treatment_module.R
\name{social_2nd_line_treatment_module}
\alias{social_2nd_line_treatment_module}
\title{Title}
\usage{
social_2nd_line_treatment_module(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
