% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_act_type_het.R
\name{social_act_type_het}
\alias{social_act_type_het}
\title{Title}
\usage{
social_act_type_het(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
