% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_coital_acts.R
\name{social_coital_acts}
\alias{social_coital_acts}
\title{Calculate number of coital acts per couple}
\usage{
social_coital_acts(dat, at)
}
\arguments{
\item{dat}{master data object}

\item{at}{timestep}
}
\value{
'dat' object with number of coital acts per couple added to 'social_discord_edgelist_df' table
}
\description{
Calculate number of coital acts per couple
}
\details{
Takes the data.frame, discord_edgelist_df, returned from social_discord_edgelist_df and calculates the number of acts per partnership for timestep then expands data.frame so each row represents single sex act if couples have more than one act per timestep; eliminates couples with zero acts for timestep.
}
\examples{
dat <- social_coital_acts(dat,at)
}
