% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_condom_use.R
\name{social_condom_use}
\alias{social_condom_use}
\title{Determine condom use per coital acts}
\usage{
social_condom_use(dat, at)
}
\arguments{
\item{dat}{master data object}

\item{at}{timestep}
}
\value{
'dat' object with status of condom use (0/1) added to 'discord_edgelist_df' table
}
\description{
Determine condom use per coital acts
}
\details{
Subfunction in 'social_coital_acts_module'.Assigns condom use (0/1) to each row (analagous to act) in discord_edgelist_df (table used to calculate transmission probabilities).
}
\examples{
dat <-social_condom_use(dat,at)
}
