% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_discord_edgelist.R
\name{social_discord_edgelist}
\alias{social_discord_edgelist}
\title{Title}
\usage{
social_discord_edgelist(el, status_vec, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here

}
