% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_generic_attribute_transition.R
\name{social_generic_attribute_transition}
\alias{social_generic_attribute_transition}
\title{Title}
\usage{
social_generic_attribute_transition(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
