% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_increased_testing_defined_groups.R
\name{social_increased_testing_defined_groups}
\alias{social_increased_testing_defined_groups}
\title{Title}
\usage{
social_increased_testing_defined_groups(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
