% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/social_increased_testing_selected_groups_module.R,
%   R/social_increased_testing_selected_groups_module_in_progress.R
\name{social_increased_testing_selected_groups_module}
\alias{social_increased_testing_selected_groups_module}
\title{Title}
\usage{
social_increased_testing_selected_groups_module(dat, at)

social_increased_testing_selected_groups_module(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.

return value here.
}
\description{
Description

Description
}
\details{
Additional details here

Additional details here
}
\examples{
example function call here
example function call here
}
