% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_role_msm.R
\name{social_role_msm}
\alias{social_role_msm}
\title{Determine sex role of agent for MSM models.}
\usage{
social_role_msm(dat, at)
}
\arguments{
\item{dat}{master data object}

\item{at}{timestep}
}
\value{
'dat' object with sex role for agents in MSM models appended to 'social_discord_edgelist_df' table
}
\description{
Determine sex role of agent for MSM models.
}
\details{
Subfunction in 'social_coital_acts_module'. Determines coital acts role for msm, any agents are other than 'versatile';  also determines if 'flipping' occurs per V-V couple, if so, adds row to discord_edgelist_df and is treated as new act.
}
\examples{
dat <- social_role_msm(dat,at)

}
