% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_set_msm_role_on_nw.R
\name{social_set_msm_role_on_nw}
\alias{social_set_msm_role_on_nw}
\title{Title}
\usage{
social_set_msm_role_on_nw(params, init_nw)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
