% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/social_testing_diagnosis_with_enhancements_module_.R
\name{social_testing_diagnosis_with_enhancements_module}
\alias{social_testing_diagnosis_with_enhancements_module}
\title{Title}
\usage{
social_testing_diagnosis_with_enhancements_module(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
