% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_treatment_prep.R
\name{social_treatement_prep}
\alias{social_treatement_prep}
\title{Title}
\usage{
social_treatement_prep(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
