% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_treatment_prep_SES.R
\name{social_treatement_prep_SES}
\alias{social_treatement_prep_SES}
\title{PrEP module}
\usage{
social_treatement_prep_SES(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
PrEP module parameterized with MSM values
}
\details{
Additional details here
}
\examples{
example function call here
}
