% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_viral_load_testing_module.R
\name{social_viral_load_testing_module}
\alias{social_viral_load_testing_module}
\title{Title}
\usage{
social_viral_load_testing_module(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
