% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_end_epidemic_values.R
\name{summary_end_epidemic_values}
\alias{summary_end_epidemic_values}
\title{Title}
\usage{
summary_end_epidemic_values(model, var, nyears)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
