% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_evo_newick.R
\name{summary_evo_newick}
\alias{summary_evo_newick}
\title{Title}
\usage{
summary_evo_newick(model, path = getwd(), name = NULL, plot = TRUE)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
