% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_partner_list.R
\name{summary_partner_list}
\alias{summary_partner_list}
\title{Title}
\usage{
summary_partner_list(dat)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
