% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_popsumm_fxns.R
\name{summary_popsumm_fxns}
\alias{summary_popsumm_fxns}
\title{Title}
\usage{
summary_popsumm_fxns(
  generic_nodal_att_values = NULL,
  aim3,
  fast_el,
  params = NULL
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
