% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_viral_lineage.R
\name{summary_viral_lineage}
\alias{summary_viral_lineage}
\title{Title}
\usage{
summary_viral_lineage(poplist)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here

}
