% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_hughes_and_exp.R
\name{transmission_hughes_and_exp}
\alias{transmission_hughes_and_exp}
\title{Title}
\usage{
transmission_hughes_and_exp(
  dat,
  acute_phase_status,
  sti_status_sus,
  condom_use,
  msm_sus_receptive_status,
  msm_sus_insert_status,
  msm_circum_status_insert_sus,
  age_vec_sus,
  hetero_sus_female_vi,
  hetero_sus_female_ai,
  hetero_sus_male_ai,
  hetero_sus_male_circum_status,
  logV_inf,
  vacc_sens_sus,
  susceptibility
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
