% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_main_module.R
\name{transmission_main_module}
\alias{transmission_main_module}
\title{Title}
\usage{
transmission_main_module(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
