% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{treatment}
\alias{treatment}
\title{Title}
\usage{
treatment(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
