% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_dropout.R
\name{treatment_dropout}
\alias{treatment_dropout}
\title{Title}
\usage{
treatment_dropout(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
