% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{utilities_annual_mortality_conversion}
\alias{utilities_annual_mortality_conversion}
\title{Title}
\usage{
utilities_annual_mortality_conversion(
  .annual_probs,
  age_range,
  .timesteps_per_year
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
