% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{utilities_annual_prob_conversion}
\alias{utilities_annual_prob_conversion}
\title{Title}
\usage{
utilities_annual_prob_conversion(annual_probs, timesteps_per_year)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
