% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{utilities_apportion.lr}
\alias{utilities_apportion.lr}
\title{Title}
\usage{
utilities_apportion.lr(vector.length, values, proportions, shuffled = F)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.

return value here.
}
\description{
Description

Description
}
\details{
Additional details here

Additional details here
}
\examples{
example function call here
example function call here
}
