% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_cd4_helper_fxns.R
\name{viral_spvl_cat_fxn}
\alias{viral_spvl_cat_fxn}
\title{Title}
\usage{
viral_spvl_cat_fxn(spvl_vec)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here

}
