% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_update_aim3_args_list_2.R
\name{viral_update_aim3_args_list_2}
\alias{viral_update_aim3_args_list_2}
\title{Title}
\usage{
viral_update_aim3_args_list_2(dat, timeIndex, ind)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
