% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_cd4_update_daily_with_age.R
\name{viral_update_cd4_daily_with_age}
\alias{viral_update_cd4_daily_with_age}
\title{Title}
\usage{
viral_update_cd4_daily_with_age(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
