% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_cd4_udate_initial_pop.R
\name{viral_update_cd4_intial_pop}
\alias{viral_update_cd4_intial_pop}
\title{Title}
\usage{
viral_update_cd4_intial_pop(dat)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
