% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_update_gamma_treated.R
\name{viral_update_gamma_treated}
\alias{viral_update_gamma_treated}
\title{Title}
\usage{
viral_update_gamma_treated(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here

}
