% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital_births_bookkeeping_misc.R
\name{vital_births_bookkeeping_misc}
\alias{vital_births_bookkeeping_misc}
\title{Title}
\usage{
vital_births_bookkeeping_misc(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
