% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital_death_aged_out.R
\name{vital_death_aged_out}
\alias{vital_death_aged_out}
\title{Title}
\usage{
vital_death_aged_out(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here

}
