% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital_initial_age_dist.R
\name{vital_initial_age_dist}
\alias{vital_initial_age_dist}
\title{Title}
\usage{
vital_initial_age_dist(age.range, popsize, age_dist)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
